# api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SpecGen API",
    "version": "1.0.0",
    "description": "API documentation for SpecGen - a speculative fiction generator",
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "schemas": {
...
      "Parameter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the parameter",
            "example": "science-fiction-technology-level"
          },
          "name": {
            "type": "string",
            "description": "Display name of the parameter",
            "example": "Technology Level"
          },
          "description": {
            "type": "string",
            "description": "Optional description of the parameter",
            "example": "The level of technological advancement in the story"
          },
          "visibility": {
            "type": "string",
            "description": "Visibility level of this parameter",
            "enum": [
              "Basic",
              "Advanced"
            ],
            "example": "Basic"
          },
          "type": {
            "type": "string",
            "description": "Type of the parameter input",
            "enum": [
              "Dropdown",
              "Slider",
              "Toggle Switch",
              "Radio Buttons",
              "Checkbox"
            ],
            "example": "Dropdown"
          },
          "categoryId": {
            "type": "string",
            "description": "ID of the category this parameter belongs to",
            "example": "science-fiction"
          },
          "values": {
            "oneOf": [
              {
                "type": "array",
                "description": "For Dropdown, Radio Buttons, and Checkbox types",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "near-future"
                    },
                    "label": {
                      "type": "string",
                      "example": "Near Future"
                    }
                  }
                },
                "example": [
                  {
                    "id": "near-future",
                    "label": "Near Future"
                  },
                  {
                    "id": "advanced",
                    "label": "Advanced"
                  }
                ]
              },
              {
                "type": "object",
                "description": "For Toggle Switch type",
                "properties": {
                  "on": {
                    "type": "string",
                    "example": "Yes"
                  },
                  "off": {
                    "type": "string",
                    "example": "No"
                  }
                },
                "example": {
                  "on": "Yes",
                  "off": "No"
                }
              }
            ]
          },
          "config": {
            "type": "object",
            "description": "Configuration for Slider type",
            "properties": {
              "min": {
                "type": "number",
                "example": 0
              },
              "max": {
                "type": "number",
                "example": 100
              },
              "step": {
                "type": "number",
                "example": 1
              }
            },
            "example": {
              "min": 0,
              "max": 100,
              "step": 1
            }
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "categoryId"
        ]
      },
...
    },
    "responses": {
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Error message details"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/parameters": {
      "get": {
        "summary": "Get all parameters or filter by categoryId",
        "tags": [
          "Parameters"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "categoryId",
            "schema": {
              "type": "string"
            },
            "description": "Optional category ID to filter parameters"
          }
        ],
        "responses": {
          "200": {
            "description": "List of parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Parameter"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "summary": "Create a new parameter",
        "tags": [
          "Parameters"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "type",
                  "categoryId"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Character Type"
                  },
                  "description": {
                    "type": "string",
                    "example": "The type of main character in the story"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "Dropdown",
                      "Slider",
                      "Toggle Switch",
                      "Radio Buttons",
                      "Checkbox"
                    ],
                    "example": "Dropdown"
                  },
                  "visibility": {
                    "type": "string",
                    "enum": [
                      "Basic",
                      "Advanced"
                    ],
                    "default": "Basic",
                    "example": "Basic"
                  },
                  "categoryId": {
                    "type": "string",
                    "example": "science-fiction"
                  },
                  "required": {
                    "type": "boolean",
                    "description": "Whether this parameter is required",
                    "default": false,
                    "example": true
                  },
                  "values": {
                    "oneOf": [
                      {
                        "type": "array",
                        "description": "For Dropdown, Radio Buttons, and Checkbox types",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "example": "hero"
                            },
                            "label": {
                              "type": "string",
                              "example": "Hero"
                            }
                          }
                        },
                        "example": [
                          {
                            "label": "Hero"
                          },
                          {
                            "label": "Villain"
                          },
                          {
                            "label": "Sidekick"
                          }
                        ]
                      },
                      {
                        "type": "object",
                        "description": "For Toggle Switch type",
                        "properties": {
                          "on": {
                            "type": "string",
                            "example": "Yes"
                          },
                          "off": {
                            "type": "string",
                            "example": "No"
                          }
                        },
                        "example": {
                          "on": "Yes",
                          "off": "No"
                        }
                      }
                    ]
                  },
                  "config": {
                    "type": "object",
                    "description": "For Slider type",
                    "properties": {
                      "min": {
                        "type": "number",
                        "example": 100
                      },
                      "max": {
                        "type": "number",
                        "example": 10000
                      },
                      "step": {
                        "type": "number",
                        "example": 100
                      }
                    },
                    "example": {
                      "min": 100,
                      "max": 10000,
                      "step": 100
                    }
                  }
                }
              },
              "examples": {
                "dropdown": {
                  "summary": "Example Dropdown Parameter",
                  "value": {
                    "name": "Character Type",
                    "type": "Dropdown",
                    "visibility": "Basic",
                    "categoryId": "science-fiction",
                    "required": true,
                    "values": [
                      {
                        "label": "Hero"
                      },
                      {
                        "label": "Villain"
                      },
                      {
                        "label": "Sidekick"
                      }
                    ]
                  }
                },
                "slider": {
                  "summary": "Example Slider Parameter",
                  "value": {
                    "name": "Story Length",
                    "type": "Slider",
                    "visibility": "Basic",
                    "categoryId": "science-fiction",
                    "required": false,
                    "config": {
                      "min": 100,
                      "max": 10000,
                      "step": 100
                    }
                  }
                },
                "toggle": {
                  "summary": "Example Toggle Parameter",
                  "value": {
                    "name": "Happy Ending",
                    "type": "Toggle Switch",
                    "visibility": "Basic",
                    "categoryId": "science-fiction",
                    "required": false,
                    "values": {
                      "on": "Yes",
                      "off": "No"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Parameter created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Parameter"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/parameters/{id}": {
      "get": {
        "summary": "Get a parameter by ID",
        "tags": [
          "Parameters"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Parameter ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Parameter details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Parameter"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "summary": "Update a parameter",
        "tags": [
          "Parameters"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Parameter ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Updated Character Type"
                  },
                  "description": {
                    "type": "string",
                    "example": "Updated description"
                  },
                  "visibility": {
                    "type": "string",
                    "enum": [
                      "Basic",
                      "Advanced"
                    ],
                    "example": "Advanced"
                  },
                  "required": {
                    "type": "boolean",
                    "example": true
                  },
                  "values": {
                    "oneOf": [
                      {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "example": "updated-hero"
                            },
                            "label": {
                              "type": "string",
                              "example": "Updated Hero"
                            }
                          }
                        },
                        "example": [
                          {
                            "label": "Updated Hero"
                          },
                          {
                            "label": "Updated Villain"
                          }
                        ]
                      },
                      {
                        "type": "object",
                        "properties": {
                          "on": {
                            "type": "string",
                            "example": "Updated Yes"
                          },
                          "off": {
                            "type": "string",
                            "example": "Updated No"
                          }
                        },
                        "example": {
                          "on": "Updated Yes",
                          "off": "Updated No"
                        }
                      }
                    ]
                  },
                  "config": {
                    "type": "object",
                    "properties": {
                      "min": {
                        "type": "number",
                        "example": 200
                      },
                      "max": {
                        "type": "number",
                        "example": 20000
                      },
                      "step": {
                        "type": "number",
                        "example": 200
                      }
                    },
                    "example": {
                      "min": 200,
                      "max": 20000,
                      "step": 200
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parameter updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Parameter"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "summary": "Delete a parameter",
        "tags": [
          "Parameters"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Parameter ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Parameter deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "deletedParameter": {
                          "$ref": "#/components/schemas/Parameter"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
...
}